rename to bbox_coder
