+++ configs/_base_/models/cascade_mask_rcnn_r50_fpn.py
@@ -23,8 +23,8 @@ model = dict(
         anchor_scales=[8],
         anchor_ratios=[0.5, 1.0, 2.0],
         anchor_strides=[4, 8, 16, 32, 64],
-        coder=dict(
-            type='DeltaCoder',
+        bbox_coder=dict(
+            type='DeltaXYWHBBoxCoder',
             target_means=[.0, .0, .0, .0],
             target_stds=[1.0, 1.0, 1.0, 1.0]),
         loss_cls=dict(
@@ -46,8 +46,8 @@ model = dict(
                 fc_out_channels=1024,
                 roi_feat_size=7,
                 num_classes=80,
-                coder=dict(
-                    type='DeltaCoder',
+                bbox_coder=dict(
+                    type='DeltaXYWHBBoxCoder',
                     target_means=[0., 0., 0., 0.],
                     target_stds=[0.1, 0.1, 0.2, 0.2]),
                 reg_class_agnostic=True,
@@ -63,8 +63,8 @@ model = dict(
                 fc_out_channels=1024,
                 roi_feat_size=7,
                 num_classes=80,
-                coder=dict(
-                    type='DeltaCoder',
+                bbox_coder=dict(
+                    type='DeltaXYWHBBoxCoder',
                     target_means=[0., 0., 0., 0.],
                     target_stds=[0.05, 0.05, 0.1, 0.1]),
                 reg_class_agnostic=True,
@@ -80,8 +80,8 @@ model = dict(
                 fc_out_channels=1024,
                 roi_feat_size=7,
                 num_classes=80,
-                coder=dict(
-                    type='DeltaCoder',
+                bbox_coder=dict(
+                    type='DeltaXYWHBBoxCoder',
                     target_means=[0., 0., 0., 0.],
                     target_stds=[0.033, 0.033, 0.067, 0.067]),
                 reg_class_agnostic=True,
