+++ configs/_base_/models/fast_rcnn_r50_fpn.py
@@ -29,8 +29,8 @@ model = dict(
             fc_out_channels=1024,
             roi_feat_size=7,
             num_classes=80,
-            coder=dict(
-                type='DeltaCoder',
+            bbox_coder=dict(
+                type='DeltaXYWHBBoxCoder',
                 target_means=[0., 0., 0., 0.],
                 target_stds=[0.1, 0.1, 0.2, 0.2]),
             reg_class_agnostic=False,
