+++ configs/_base_/models/ssd300.py
@@ -21,8 +21,8 @@ model = dict(
         anchor_strides=(8, 16, 32, 64, 100, 300),
         basesize_ratio_range=(0.15, 0.9),
         anchor_ratios=([2], [2, 3], [2, 3], [2, 3], [2], [2]),
-        coder=dict(
-            type='DeltaCoder',
+        bbox_coder=dict(
+            type='DeltaXYWHBBoxCoder',
             target_means=[.0, .0, .0, .0],
             target_stds=[0.1, 0.1, 0.2, 0.2])))
 cudnn_benchmark = True
