+++ configs/atss/atss_r50_fpn.py
@@ -28,8 +28,8 @@ model = dict(
         scales_per_octave=1,
         anchor_ratios=[1.0],
         anchor_strides=[8, 16, 32, 64, 128],
-        coder=dict(
-            type='DeltaCoder',
+        bbox_coder=dict(
+            type='DeltaXYWHBBoxCoder',
             target_means=[.0, .0, .0, .0],
             target_stds=[0.1, 0.1, 0.2, 0.2]),
         loss_cls=dict(
