+++ configs/free_anchor/retinanet_free_anchor_r50_fpn_1x_coco.py
@@ -11,8 +11,8 @@ model = dict(
         scales_per_octave=3,
         anchor_ratios=[0.5, 1.0, 2.0],
         anchor_strides=[8, 16, 32, 64, 128],
-        coder=dict(
-            type='DeltaCoder',
+        bbox_coder=dict(
+            type='DeltaXYWHBBoxCoder',
             target_means=[.0, .0, .0, .0],
             target_stds=[0.1, 0.1, 0.2, 0.2]),
         loss_bbox=dict(type='SmoothL1Loss', beta=0.11, loss_weight=0.75)))
