+++ configs/guided_anchoring/ga_retinanet_r50_caffe_fpn_1x_coco.py
@@ -13,11 +13,11 @@ model = dict(
         anchor_strides=[8, 16, 32, 64, 128],
         anchor_base_sizes=None,
         anchor_coder=dict(
-            type='DeltaCoder',
+            type='DeltaXYWHBBoxCoder',
             target_means=[.0, .0, .0, .0],
             target_stds=[1.0, 1.0, 1.0, 1.0]),
-        coder=dict(
-            type='DeltaCoder',
+        bbox_coder=dict(
+            type='DeltaXYWHBBoxCoder',
             target_means=[.0, .0, .0, .0],
             target_stds=[1.0, 1.0, 1.0, 1.0]),
         loc_filter_thr=0.01,
