+++ configs/guided_anchoring/ga_rpn_r50_fpn_1x_coco.py
@@ -11,11 +11,11 @@ model = dict(
         anchor_strides=[4, 8, 16, 32, 64],
         anchor_base_sizes=None,
         anchor_coder=dict(
-            type='DeltaCoder',
+            type='DeltaXYWHBBoxCoder',
             target_means=[.0, .0, .0, .0],
             target_stds=[0.07, 0.07, 0.14, 0.14]),
-        coder=dict(
-            type='DeltaCoder',
+        bbox_coder=dict(
+            type='DeltaXYWHBBoxCoder',
             target_means=[.0, .0, .0, .0],
             target_stds=[0.07, 0.07, 0.11, 0.11]),
         loc_filter_thr=0.01,
