+++ mmdet/core/bbox/__init__.py
@@ -1,22 +1,21 @@
 from .assigners import AssignResult, BaseAssigner, MaxIoUAssigner
-from .coder import BaseCoder, DeltaCoder, PseudoCoder
+from .coder import BaseBBoxCoder, DeltaXYWHBBoxCoder, PseudoBBoxCoder
 from .iou_calculators import BboxOverlaps2D, bbox_overlaps
 from .samplers import (BaseSampler, CombinedSampler,
                        InstanceBalancedPosSampler, IoUBalancedNegSampler,
                        PseudoSampler, RandomSampler, SamplingResult)
-from .transforms import (bbox2delta, bbox2result, bbox2roi, bbox_flip,
-                         bbox_mapping, bbox_mapping_back, delta2bbox,
-                         distance2bbox, roi2bbox)
+from .transforms import (bbox2result, bbox2roi, bbox_flip, bbox_mapping,
+                         bbox_mapping_back, distance2bbox, roi2bbox)
 
 from .builder import (  # isort:skip, avoid recursive imports
-    build_assigner, build_sampler, build_coder)
+    build_assigner, build_sampler, build_bbox_coder)
 
 __all__ = [
     'bbox_overlaps', 'BboxOverlaps2D', 'BaseAssigner', 'MaxIoUAssigner',
     'AssignResult', 'BaseSampler', 'PseudoSampler', 'RandomSampler',
     'InstanceBalancedPosSampler', 'IoUBalancedNegSampler', 'CombinedSampler',
-    'SamplingResult', 'build_assigner', 'build_sampler', 'bbox2delta',
-    'delta2bbox', 'bbox_flip', 'bbox_mapping', 'bbox_mapping_back', 'bbox2roi',
-    'roi2bbox', 'bbox2result', 'distance2bbox', 'build_coder', 'BaseCoder',
-    'PseudoCoder', 'DeltaCoder'
+    'SamplingResult', 'build_assigner', 'build_sampler', 'bbox_flip',
+    'bbox_mapping', 'bbox_mapping_back', 'bbox2roi', 'roi2bbox', 'bbox2result',
+    'distance2bbox', 'build_bbox_coder', 'BaseBBoxCoder', 'PseudoBBoxCoder',
+    'DeltaXYWHBBoxCoder'
 ]
