+++ mmdet/core/bbox/builder.py
@@ -1,7 +1,7 @@
 from mmdet.utils import build_from_cfg
 from .assigners import BaseAssigner
-from .coder import BaseCoder
-from .registry import BBOX_ASSIGNERS, BBOX_CODER, BBOX_SAMPLERS
+from .coder import BaseBBoxCoder
+from .registry import BBOX_ASSIGNERS, BBOX_CODERS, BBOX_SAMPLERS
 from .samplers import BaseSampler
 
 
@@ -17,7 +17,7 @@ def build_sampler(cfg, **default_args):
     return build_from_cfg(cfg, BBOX_SAMPLERS, default_args)
 
 
-def build_coder(cfg, **default_args):
-    if isinstance(cfg, BaseCoder):
+def build_bbox_coder(cfg, **default_args):
+    if isinstance(cfg, BaseBBoxCoder):
         return cfg
-    return build_from_cfg(cfg, BBOX_CODER, default_args)
+    return build_from_cfg(cfg, BBOX_CODERS, default_args)
