+++ mmdet/core/bbox/registry.py
@@ -2,4 +2,4 @@ from mmdet.utils import Registry
 
 BBOX_ASSIGNERS = Registry('bbox_assigner')
 BBOX_SAMPLERS = Registry('bbox_sampler')
-BBOX_CODER = Registry('bbox_coder')
+BBOX_CODERS = Registry('bbox_coder')
